requirements.txt
basicauth==0.3
boto==2.15.0
Django==1.7.4
django-apptemplates==0.0.1
django-colorful==1.0.1
httplib2==0.9
MySQL-python==1.2.5
oauth2==1.5.211
Pillow==2.7.0
py-pretty==1
python-instagram==1.3.0
python-memcached==1.53
pytz==2014.10
requests==2.5.1
simplejson==3.6.5
six==1.9.0
git+https://github.com/chrsbats/kvstore.git#egg=kvstore
wsgiref==0.1.2
psycopg2==2.6
whitenoise==1.0.6
django-storages==1.1.8
